requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.1
django-cors-headers==3.4.0
django-debug-toolbar==3.1.1
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==1.7.0
install==1.3.4
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
openpyxl==3.0.5
Pillow==7.2.0
psycopg2-binary==2.8.5
pyaml==20.4.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
xlrd==2.0.1
zipp==3.1.0
